update register useCases
